requirements.txt
aiofiles==24.1.0
backports.tarfile==1.2.0
fastapi==0.116.1
httptools==0.6.4
importlib-metadata==8.0.0
jaraco.collections==5.1.0
jinja2==3.1.6
packaging==24.2
pip-chill==1.0.3
platformdirs==4.2.2
python-dotenv==1.1.1
python-multipart==0.0.20
pyyaml==6.0.2
tomli==2.0.1
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
itsdangerous==2.2.0
wordfreq==3.1.1
sqlmodel==0.0.21
requests>=2.31,<3.0
psycopg[binary]>=3.2,<4
alembic>=1.13
